// patchnotes-web/src/api/generated/model/packageDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * PatchNotes.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PackageDto {
  id: string;
  name: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  npmName?: string | null;
  githubOwner: string;
  githubRepo: string;
  /** @nullable */
  tagPrefix?: string | null;
  /** @nullable */
  lastFetchedAt?: string | null;
  createdAt?: string;
}