// patchnotes-web/src/api/generated/model/releaseDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * PatchNotes.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ReleasePackageDto } from './releasePackageDto';

export interface ReleaseDto {
  id: string;
  tag: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  publishedAt?: string;
  fetchedAt?: string;
  package: ReleasePackageDto;
}